Helps editing when copying from PDFs
When I was copying from certain PDFs, occasionally spaces were replaced by the newline character.
I made this program to open a file (which contained the uncorrected text) and print a corrected form of the text to standard output. Then, I used the terminal to direct that to a file.